move tool logic to __init__.py
